pyproject.toml
[project]
name = "snakecharmer"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28,<3",
    "pyyaml>=6.0,<7",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
